// ecoyaan-checkout/types/cart.ts
export interface CartItem {
  product_id: number;
  product_name: string;
  product_price: number;
  quantity: number;
  image: string;
}

export interface CartResponse {
  cartItems: CartItem[];
  shipping_fee: number;
  discount_applied: number;
}

export interface ShippingAddress {
  fullName: string;
  email: string;
  phone: string;
  pinCode: string;
  city: string;
  state: string;
}

export interface OrderSummary {
  subtotal: number;
  shippingFee: number;
  discount: number;
  grandTotal: number;
}